Finished a command I wanted to make.
